mptcp: reset 'first' and ack_hint on subflow close
Git-commit: c8fe62f0768cc9378103fc89fb96804645f527c9
Message-Id: <c8fe62f0768cc9378103fc89fb96804645f527c9.1637074015.git.geliang.tang@suse.com>
References: bsn#2
Patch-mainline: v5.12-rc3

Just like with last_snd, we have to NULL 'first' on subflow close.

ack_hint isn't strictly required (its never dereferenced), but better to
clear this explicitly as well instead of making it an exception.

msk->first is dereferenced unconditionally at accept time, but
at that point the ssk is not on the conn_list yet -- this means
worker can't see it when iterating the conn_list.

Reported-by: Paolo Abeni <pabeni@redhat.com>
Reviewed-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Signed-off-by: Florian Westphal <fw@strlen.de>
Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.suse/v5.12-rc3-mptcp-reset-first-and-ack_hint-on-subflow-close.patch